backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "publisher-ai-assistant"
version = "0.1.0"
description = "LangChain-based AI assistant for publishing workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "pgvector>=0.3.3",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "PyJWT>=2.8.0",
  "typer>=0.12.3",
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-text-splitters>=0.3.0",
  "mcp>=1.27.0",
  "pypdf>=4.2.0",
  "pymupdf>=1.24.0",
  "pymupdf4llm>=0.0.17",
  "python-docx>=1.1.2",
  "elasticsearch>=8.15.0",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "pytest-cov>=5.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
addopts = "--basetemp=data/tmp/pytest"
testpaths = ["tests"]
pythonpath = ["."]
